Wrapping CSV State Machine

+++ src/execution/operator/csv_scanner/sniffer/type_detection.cpp
@@ -151,8 +151,8 @@ struct SniffValue {
 		}
 		machine.pre_previous_state = machine.previous_state;
 		machine.previous_state = machine.state;
-		machine.state = static_cast<CSVState>(
-		    machine.transition_array[static_cast<uint8_t>(machine.state)][static_cast<uint8_t>(current_char)]);
+		machine.state =
+		    static_cast<CSVState>(machine.transition_array[machine.state][static_cast<uint8_t>(current_char)]);
 
 		bool carriage_return = machine.previous_state == CSVState::CARRIAGE_RETURN;
 		if (machine.previous_state == CSVState::DELIMITER ||
